requirements.txt
tensorboard
torchtext
numpy
pandas
utils>=1.0.0
sacrebleu
nltk
tqdm
flair
scipy==1.10.1